+++ pkg/artifactory/resource/webhook/resource_artifactory_custom_webhook_release_bundle.go
@@ -1,3 +1,17 @@
+// Copyright (c) JFrog Ltd. (2025)
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package webhook
 
 import (
@@ -17,9 +31,10 @@ func NewArtifactoryReleaseBundleCustomWebhookResource() resource.Resource {
 	return &ReleaseBundleCustomWebhookResource{
 		CustomWebhookResource: CustomWebhookResource{
 			WebhookResource: WebhookResource{
-				TypeName:    fmt.Sprintf("artifactory_%s_custom_webhook", ArtifactoryReleaseBundleDomain),
-				Domain:      ArtifactoryReleaseBundleDomain,
-				Description: "Provides an Artifactory webhook resource. This can be used to register and manage Artifactory webhook subscription which enables you to be notified or notify other users when such events take place in Artifactory.:",
+				TypeName: fmt.Sprintf("artifactory_%s_custom_webhook", ArtifactoryReleaseBundleDomain),
+				Domain:   ArtifactoryReleaseBundleDomain,
+				Description: "Provides an Artifactory webhook resource. This can be used to register and manage Artifactory webhook subscription which enables you to be notified or notify other users when such events take place in Artifactory.\n\n" +
+					"!>This resource is being deprecated and replaced by `artifactory_destination_custom_webhook` resource.",
 			},
 		},
 	}
@@ -53,10 +68,9 @@ func NewReleaseBundleCustomWebhookResource() resource.Resource {
 	return &ReleaseBundleCustomWebhookResource{
 		CustomWebhookResource: CustomWebhookResource{
 			WebhookResource: WebhookResource{
-				TypeName: fmt.Sprintf("artifactory_%s_custom_webhook", ReleaseBundleDomain),
-				Domain:   ReleaseBundleDomain,
-				Description: "Provides an Artifactory webhook resource. This can be used to register and manage Artifactory webhook subscription which enables you to be notified or notify other users when such events take place in Artifactory.\n\n" +
-					"!>This resource is being deprecated and replaced by `artifactory_destination_webhook` resource.",
+				TypeName:    fmt.Sprintf("artifactory_%s_custom_webhook", ReleaseBundleDomain),
+				Domain:      ReleaseBundleDomain,
+				Description: "Provides an Artifactory webhook resource. This can be used to register and manage Artifactory webhook subscription which enables you to be notified or notify other users when such events take place in Artifactory.:",
 			},
 		},
 	}
@@ -76,7 +90,7 @@ func (r *ReleaseBundleCustomWebhookResource) Metadata(ctx context.Context, req r
 
 func (r *ReleaseBundleCustomWebhookResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = r.CreateSchema(r.Domain, &releaseBundleCriteriaBlock)
-	if r.Domain == ReleaseBundleDomain {
+	if r.Domain == ArtifactoryReleaseBundleDomain {
 		resp.Schema.DeprecationMessage = "This resource is being deprecated and replaced by artifactory_destination_webhook resource"
 	}
 }